split core-schema.ts into domain-aligned schema files
